Make /vt/logs default destination for logs.
